(#1065) Applied new naming convention in text package
