+++ src/main/java/org/cactoos/text/package-info.java
@@ -26,9 +26,9 @@
  * Text.
  *
  * @since 0.1
- * @todo #1005:30min Continue applying the new class naming convention:
- *  avoid compound names for decorators. Continue renaming classes implementing
- *  the {@link org.cactoos.Text}. More details you can find here
+ * @todo #1065:30min Continue applying the new class naming convention:
+ *  avoid compound names for decorators. Continue renaming of classes
+ *  implementing the {@link org.cactoos.Text}. More details you can find here
  *  https://github.com/yegor256/cactoos/issues/913#issuecomment-402332247.
  */
 package org.cactoos.text;
